+++ CHANGES.rst
@@ -7,6 +7,9 @@ CHANGES
 - The tests are not compatible with (yanked) ``Chameleon`` 4.3.0, thus not
   allowing to use this version.
 
+- Update translation handling to be compatible with Chameleon 4.3+: Only
+  translate message ids.
+
 
 4.0 (2023-02-09)
 ================